1.3-pre0 commit * Removed warnings * ib-verbs/disconnect (all ibv_* calls implemented) * in setvolume sending one more dictionary argument "ERROR" with proper error string * made default aggregate-size in 'write-behind' xlator as 128k from 1M. (to get more performance in case of ib-verbs).

+++ libglusterfs/src/dict.c
@@ -146,7 +146,7 @@ dict_set (dict_t *this,
   char key_free = 0;
 
   if (!key) {
-    asprintf (&key, "ref:0x%x", value);
+    asprintf (&key, "ref:0x%x", (uint32_t)value);
     key_free = 1;
   }
